moved enums to separate classes
