+++ InputDialog/IDButton.cs
@@ -1,0 +1,8 @@
+﻿namespace InputDialog;
+public enum IDButton
+{
+    Ok,
+    OkCancel,
+    YesNo,
+    YesNoCancel
+}
