+++ InputDialog/IDIcon.cs
@@ -1,0 +1,9 @@
+﻿namespace InputDialog;
+public enum IDIcon
+{
+    Error,
+    Exclamation,
+    Information,
+    Question,
+    Nothing
+}
