+++ InputDialog/IDType.cs
@@ -1,0 +1,8 @@
+﻿namespace InputDialog;
+public enum IDType
+{
+    ComboBox,
+    TextBox,
+    MsgBox,
+    Numeric
+}
