+++ InputDialogUsageSample/Main.cs
@@ -4,7 +4,7 @@ namespace InputDialogUsageSample;
 
 public partial class Main : Form
 {
-    InputDialog.InputDialog.IDIcon selectedIcon = InputDialog.InputDialog.IDIcon.Question;
+    InputDialog.IDIcon selectedIcon = InputDialog.IDIcon.Question;
 
     public Main()
     {
@@ -21,8 +21,8 @@ public partial class Main : Form
             txt,
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox);
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox);
         if (rslt.DialogResult == DialogResult.OK)
             ShowResult(rslt.ResultText);
     }
@@ -33,8 +33,8 @@ public partial class Main : Form
             "Enter text",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox,
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox,
             defaultText: "Default Text");
         if (rslt.DialogResult == DialogResult.OK)
             ShowResult(rslt.ResultText);
@@ -46,8 +46,8 @@ public partial class Main : Form
             "Enter text",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox,
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox,
             buttonTexts: new ButtonTexts { OKText = "Do It long button text here" });
         if (rslt.DialogResult == DialogResult.OK)
             ShowResult(rslt.ResultText);
@@ -59,7 +59,7 @@ public partial class Main : Form
             "This is the message.",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel);
+            InputDialog.IDButton.OkCancel);
     }
 
     private void btnComboBoxWithError_Click(object sender, EventArgs e)
@@ -68,8 +68,8 @@ public partial class Main : Form
             "This is the message.",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            type: InputDialog.InputDialog.IDType.ComboBox);
+            InputDialog.IDButton.OkCancel,
+            type: InputDialog.IDType.ComboBox);
         if (rslt.DialogResult == DialogResult.OK)
             ShowResult(rslt.ResultText);
     }
@@ -82,8 +82,8 @@ public partial class Main : Form
                 "Enter text",
                 "Title",
                 selectedIcon,
-                InputDialog.InputDialog.IDButton.OkCancel,
-                InputDialog.InputDialog.IDType.TextBox,
+                InputDialog.IDButton.OkCancel,
+                InputDialog.IDType.TextBox,
                 formFont: new Font(fd.Font.FontFamily, fd.Font.Size, fd.Font.Style));
             if (rslt.DialogResult == DialogResult.OK)
                 ShowResult(rslt.ResultText);
@@ -96,8 +96,8 @@ public partial class Main : Form
             "Enter Text",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox,
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox,
             backgroundColor: Color.AliceBlue);
         if (rslt.DialogResult == DialogResult.OK)
             ShowResult(rslt.ResultText);
@@ -109,8 +109,8 @@ public partial class Main : Form
             "Enter Text",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox,
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox,
             foregroundColor: Color.Red);
         if (rslt.DialogResult == DialogResult.OK)
             ShowResult(rslt.ResultText);
@@ -122,8 +122,8 @@ public partial class Main : Form
             "Enter Text",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox,
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox,
             foregroundColor: Color.White,
             backgroundColor: Color.DarkBlue);
         if (rslt.DialogResult == DialogResult.OK)
@@ -136,8 +136,8 @@ public partial class Main : Form
             "Enter Text",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            InputDialog.InputDialog.IDType.TextBox,
+            InputDialog.IDButton.OkCancel,
+            InputDialog.IDType.TextBox,
             foregroundColor: Color.White,
             formFont: new Font("Arial", 28, FontStyle.Bold),
             backgroundImage: Image.FromFile(@"Images\Picture.jpg"),
@@ -157,8 +157,8 @@ public partial class Main : Form
            "This is the message.",
            "Title",
            selectedIcon,
-           InputDialog.InputDialog.IDButton.OkCancel,
-           type: InputDialog.InputDialog.IDType.ComboBox,
+           InputDialog.IDButton.OkCancel,
+           type: InputDialog.IDType.ComboBox,
            listItems: new List<string> { "Item 1", "Item2", "Item 3" },
            acceptsUserInput: true);
         if (rslt.DialogResult == DialogResult.OK)
@@ -171,8 +171,8 @@ public partial class Main : Form
             "This is the message.",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            type: InputDialog.InputDialog.IDType.ComboBox,
+            InputDialog.IDButton.OkCancel,
+            type: InputDialog.IDType.ComboBox,
             listItems: new List<string> { "Item 1", "Item2", "Item 3" },
             acceptsUserInput: false);
         if (rslt.DialogResult == DialogResult.OK)
@@ -191,23 +191,23 @@ public partial class Main : Form
         switch (selectedItem)
         {
             case "Error":
-                selectedIcon = InputDialog.InputDialog.IDIcon.Error;
+                selectedIcon = IDIcon.Error;
                 mnuIconSelectorError.Checked = true;
                 break;
             case "Exclamation":
-                selectedIcon = InputDialog.InputDialog.IDIcon.Exclamation;
+                selectedIcon = IDIcon.Exclamation;
                 mnuIconSelectorExclamation.Checked = true;
                 break;
             case "Information":
-                selectedIcon = InputDialog.InputDialog.IDIcon.Information;
+                selectedIcon = IDIcon.Information;
                 mnuIconSelectorInformation.Checked = true;
                 break;
             case "Question":
-                selectedIcon = InputDialog.InputDialog.IDIcon.Question;
+                selectedIcon = IDIcon.Question;
                 mnuIconSelectorQuestion.Checked = true;
                 break;
             default:
-                selectedIcon = InputDialog.InputDialog.IDIcon.Nothing;
+                selectedIcon = IDIcon.Nothing;
                 mnuIconSelectorNone.Checked = true;
                 break;
         }
@@ -219,8 +219,8 @@ public partial class Main : Form
             "Use the arrows to select a value.",
             "Title",
             selectedIcon,
-            InputDialog.InputDialog.IDButton.OkCancel,
-            type: InputDialog.InputDialog.IDType.Numeric,
+            InputDialog.IDButton.OkCancel,
+            type: InputDialog.IDType.Numeric,
             numericProperties: new NumericProperties 
             {
                 //Try out these options that change the display
